This is just a backup, it's not working and subject to violent change!
.gitignore - Adding *.d.ts files since they are generated

lerna.json & package-lock.json - Added Lerna support

package.json - Edited it to get rid of main since we are now a monorepo.

cli/package.json - Created for command line tool for Sleeve

common-utilities & related package.json - Some common utilities

common-utilities/resource.ts - The structure for a generic resource

resourcegroup.ts - Logic for creating a resource group

resourcegroup-azure/setup.ts - First scratches at logic to set up a
directory with a resource group.

webapp-node-azure/package.json - Future home of the logic to deal with
node web apps

src/index.ts - Got rid of this since we are a mono-repo and this is in
the wrong place

tsconfig.json - I strongly suspect I don't need moduleResolution equal to
node but it seemed like a good idea to add it since we are building
node modules. I got rid of strictNullChecks and replaced it with strict
which includes strictNullChecks and other fun stuff. And I suspect
declaration is true by default but I felt like sticking it in anyway.

tsling.json - Because the holy number is 80, 80 is the holy number. :)