Got rid of Lerna
I was running into issues with TSC and Lerna not liking each other (I
think that was the problem) and I realized that for a hack Lerna is anyway
mega overkill so I pulled it out.

The only other change was moving the setup logic for resource groups into
the resource group class.