Just a backup, doesn't even work
package.json/package-lock.json - Added yargs to help with command line
processing and tmp-promise to make it easy to create temporary files for
outputting and executing powershell scripts.

cli.ts - Processing setup and deploy commands from the cli

common-utilities - Enable us to run powershell scripts, see below

resource.ts - Added setup as a static method as well as properties to
hold the resource's basename (it's friendly name) and its path. Also
added deployResource to resource so it's inherited by everybody. Also
fixed a problem with inheritence for fluent design. Need to either not
specify return value or specify this which allows for polymorphism.

resourcegroup - Made compliant with new resource design and got rid of
problem with not returning this for fluent design. Now we output text
for a powershell script. This mimics what things will look like when we
support ARM templates.

servicesToCreate, sleeveforarm - Resource reference wasn't needed

webapp-node-azure - Made output powershell and setup function and made
compliant with new resource and fluent design issue

resourcegroupTests - Hacked up, need to fix

webapp-node-azureTests - Hacked up to new design

tslint.json - We need to be able to specify function and not arrow
function in some cases because of issues with capturing the right this
pointer.