Just a backup
.gitignore - Was playing with IntelliJ because of a problem I had with
VSCode

assets/cliInit/keyVault/* - Added a default keyvault to store secrets

assets/keyvault - For people creating their own keyvaults

package-lock.json - Deleted it since we had to move to an old version of
npm anyway

IGlobalDefaults - To keep consistent on how we specify this

cli - Calling npm setup earlier so auto complete will work from the get go
in IDEs like VS Code. Also added strict to yargs so we will return an
error if we get unrecognized arguments.

common-utilities - Keep running into the pattern of needing to iterate
through all the resources in a Sleeve project so added a function

keyvault - Adding support for key vaults

resource - Adding in early npmSetup for auto complete

resourceNotResourceGroup - Created this as a template for resources that
are everything but a resource group. This acts as an intermediary class
on top of Resource. We needed this because all other resource types need
to be associated with a resource group except, of course, resource groups.

resourcegroup - Added IGlobalDefault support, also added the
testingCalculateResourceGroupName function for testing purposes but my
goal is to get rid of it and instead be able to use the output from the
sleeve deploy command to get the same data.

sleeve.cmd - So we can look like a normal CLI

webapp-node-azure - Refactored to take advantage of
resourceNotResourceGroup and put in setDefaultWebAppDNSName as an explicit
function but I hope to eventually hide it and have it triggered as part
of setting the resourceGroup value.

webapp-node-azureTest - Edited to use sleeve commands directly for setup.

testconfig.json - Including the testingRoot files which include code from
test deployments was confusing TSC to the point where it stopped working so
I needed to exclude them.