Getting the new refactoring done and storing secrets in keyvault
.gitignore - Had to move the testing files to keep VS Code from getting
confused and breaking

cli.ts & cliUtilities.ts - Refactored into two files, cli.ts which is
called only on the command line and cliUtilities.ts where most of the
code lives. This is necessary because yargs runs automatically and so
causes confusion in testing.

common-utilities - Tools to search for specific resource types

keyvaultInfrastructure - Added the ability to set secrets

resource.ts - Since Typescript doesn't support double inheritance I had
to stick some infrastructure properties here so they could be inherited.
But the good news is that they are all protected and so will be invisible
to our users.

resourceNotResourceGroup - Just using a new utility and officially
recording the resourceInEnvironment for this deployment.

webapp-node-azureInfrastructure - Adding storing secrets in key vault

testUtilities - We moved the test folder as mentioned above

webapp-node-azureTest - Dealing with change in test file location and
also calling deploy in code to make it easier to debug.

tsconfig.json - Since we moved testingRoot (with a new name) to a location
that isn't in the include we don't need it in exclude.