Publish to npm & support setting firewall rules
package.json - Upping version #

cli.ts - Adding a check if we are supposed to be logging

cliUtilities - Working around bizarre instanceof behavior in typescript.
And added function to support logging to screen and file which can be
trigger from the CLI.

common-utilities - Added an insecure function to get the local machine's
IP for when we have to set firewall rules for this machine. Sometimes
the firewall rules can take a little bit to kick in so created a retry
mechanism for commands. Function to deal with bizarre instanceof failures.

mysql-azure - Changed initialize command so you can call it multiple times.
This is easier to understand than one call overriding another.

mysql-azureInfrastructure - Support multiple mySQL initialize scripts.
Now setting temporary firewall rule to let the dev machine run the
init script and then removing the rule when done. Added logic to add
and remove firewall rules.

mysql-azureTest - Set test to prod and updated the function name for
initing mySQL