Now powershell free!
baseDeployStorageResource - We now create the name of the firewall rule
ourselves both to make sure it meets the restrictions on what characters
can be in the rule (no periods) as well as to make sure we know what
resource required the rule for auditing purposes.

cliUtilities - Forget the await so I was returning a promise

common-utilities - Added some debug statements before executing the
commands so I can see what we are waiting for when looking at the logs
during a debug run. Deleted appendErrorCheck because we don't need it now
that we got rid of powershell scripts.

keyvaultInfrastructure - Getting rid of quotes just to be clean.

mysql-azureInfrastructure - Changed setFirewallRule per previous
description. Made sure we only set up the firewall rule to let us
run config scripts when we have actual scripts to run. We also had a bug
that the finally that cleaned up the temporary firewall rule that gives
the dev machine access to production was run before we tried to actually
run the script. Got rid of quotes in commands just to be cleaner.

resourcegroupInfrastructure - Getting rid of quotes.

webapp-node-azureInfrastructure - The first change is because it seems
that the await was waiting for the query part of the conditional operator,
not the result. So used parenthesis to make sure it waits for the result.
I also completely forgot to trip the promiseGate so that those trying
to get the basedeploy object would get it! Removed unnecessary quotes from
commands.

webapp-node-azureTest - Put back the try catch around the tsc call because
it's compiling some unclear TS code for the test. So it will get errors.
Fixed a bug resulting of my change of WebappNodeAzureInfrastructure from
a default export to a non-default export. Adding some typing to the
waitAndTryAgain function because I could.