Fixing bugs
README.md - Making instructions clearer

src/cliUtilities - Added an option to delete resources before deploying,
use this to clean things up for tests.

common-utilities - Just typing the response, primarily for retryAfterFailure
so we can capture output we need to compare for tests.

mysql-azureInfrastructure - I screwed up the location of the finally so
I ended up deleting the firewall rule before using it to run a script. I
shortened the firewall rule name just, because. I also fixed a bug where
I forgot that with the new code we don't specify the firewall rule name
directly, we provide resource and IP and the firewall rule name is
synthesized from that. So we need to capture the firewall rule name
that is returned by setFirewallRule so we can delete it. I changed the
firewall rule construction in setFirewallRule so that we can identify
which resource the firewall rule was set on behalf of. And I made
setFirewallAllowAll public so I could use it for tests.

resourcegroupInfrastructure - Added logic to delete a resource group so
we can clean up the environment before running tests.

mysql-azureTest - Deleting any existing resource groups before running
the test. Now I run a simple check to see if we got the response we were
expecting from the database we set up with an init script.

webapp-node-azureTest - Increased time out because with the new logic
to delete the resource group before running the test things take longer.
Fixed a bug that caused an uncaught promise error in the ECONNREFUSED
check. Now both tests delete the resource group before running the test.