Fixes for #15 and #16
DEVREADME - Clarifying a few things that folks new to Node or Typescript
might not immediately know

cliUtilities - Fix for #16

common-utilities - Fix for #15

resource - I needed the baseName available so that in cliUtilities when I
print out the URL a web app is available at I can include the name of the
directory (aka the resource) that generated the web app. Technically the
name is in the URL but still.