Removing fs-extra-promise
We were using fs-extra-promise in a lot of places but it doesn't seem to
be maintained anymore and fs-extra now handles promises. So it's not
necessary. I also needed a function that was in fs-extra but not
fs-extra-promise so it was time to remove fs-extra-promise.

README - Clarifying our mission

fix-fs-extra-promise - I added this to fix some incorrect declarations in
the @types for fs-extra-promise but since we aren't suing fs-extra-promise
I don't need it anymore.

package-json - Removing fs-extra-promise and also tmp-promise which I got
rid of with the powershell execution function.

cli, cliUtilities, mysql-azureInfrastructure, resource,
serviceEnvironmentUtilities, webapp-node-azureInfrastructure,
mysql-azureTest, resourcegroupTest, testUtilities,
webapp-node-azureTest - Getting rid of fs-extra-promise

common-utilities - Updating for fs-extra-promise and getting rid of
runPowerShellScript since we don't do that anymore. This also meant I could
drop tmp-promise. Also added findGitRootDir as part of the solution for
#9.