Fix #9
We can now run deploy and setup commands from anywhere inside of a sleeve
project (well, o.k., not strictly true, due to issues with Node it won't
work so well if we are inside a node_modules directory)

README - We will now make a sleeve project a GIT project if it isn't
already so we need to tell people to have Git installed.

cli - Added logic on init to make a project into a git project if it isn't
already. Also added logic to print the path to a newly created resource
(via setup) so the user doesn't get confused if they ran setup in a
different directory than the root and wonders where their resource is.

cliUtilities - Added logic on setup to find the root of the project and
create the new resource under it and to return the path we created the
resource in so we can print it out for the user. Also changed
deployResources to find the root of the project from the current working
directory.