Issue #10 - Defining location during init
assets/cliInit/sleeve.js - Added setDataCenter with a place holder that
we can fill in during init with the actual chosen location.

declaration.d.ts - Got rid of tmp-promise because we don't use it anymore
(it was deleted with the powershell script function) and added
replace-in-file which we use for sleeve.js above.

package.json - adding in replace-in-file per previous

cli - Refactored init into a function in CliUtilities

cliUtilities - Moved init code here (since it was growing and I wanted to
keep cli.ts clean so I can easily read the YARG config) and added the logic
to figure out what location we want to put the project in and setting that
location in the root sleeve.js file. I also added logic to make sleeve
projects into git repo's if they aren't already.

resource - Added a list of DC names for auto-complete

resourcegroup - Added setDataCenter and dataCenter (get) so we can set
locations in init

resourcegroupInfrastructure - Added logic to set data center based on the
new dataCenter variable.