Minor tweaks
README - Didn't give a name to the sample project os everyone doesn't try
to create the same project with the same name.

cli - Changed the terminology from 'service' to 'resource' since service is
everything and resources are the components of a service. Also put in a
temporary hack around #19 and added support for specifying version.