Sleeve command line mess
So I wanted to call the command line "sleeve" but it turns out that this
causes a real problem on my Windows box. When I'm in a project directory
with a sleeve.js file when I type "sleeve" Windows will open notepad with
the sleeve.js file rather than call the sleeve command. So I had to change
the command to sleeveforarm to prevent the collision.

README - Changing the command name and adding some more explanatory text.

package.json - Using the 'brief" bin format which automatically names the
executable the same as the package, sleeveforarm.