Some fix ups
README - Nice to be polite and also not suggesting a name for the project

.npmignore - There is an annoying habit (for good reason) in NPM that they
turn .gitignore into .npmignore which causes problems when I'm trying to
create a .gitignore file in a template. So the fix is I create a .npmignore
and then during init I rename it.

package.json - Making sure things work and changed sleeve.cmd to
sleeveforarm.cmd. It actually doesn't matter (NPM puts in an indirection
so the command will be sleeveforarm even if the cmd file is foo.cmd) but it
seemed more consistent.

cliUtilities - Renaming .npmignore to .gitignore

sleeve.cmd & sleeveforarm.cmd - Changing to sleeveforarm.cmd

testUtilities - Just updating the name of the cmd file now that it has
changed