#30
webapp-node-azureinfrastructure - Retry the git clone so if it fails (which
it will because Azure isn't ready) we will try again.

webapp-node-azureTest - Just a change to let us look for other results than
"A Name". I added this for a test I later removed from this file but it
seems a useful change to the function so I am keeping the function change.