Finish touches on AI
README - Added information about the manage command and updated the sample
code to include the hooks for application insights.

index.js - I think auto dependency correlation is actually supposed to be on
but I could be wrong.

applicationInsightsInfrastructure - Did I move a space or something?

cli.ts - Added manage command

cliUtilities - Added support for AI and manage command

common-utilities - So AZ commands return empty string when there is no
result even when we expect JSON so I handled that case.

mysql-azureInfrastructure - It turns out that in a certain failure case we
can hang forever. We need to put in a proper fix (since that failure is
a potential security hole) but for now I just put in a work around.

webapp-node-azureInfrastructure - I simplified the code by waiting for
AI to return (otherwise I would have had to pick it out of
storagePromisesToWaitFor which didn't seem worth the complexity). It
shouldn't hurt perf as we already have fired off the other promises.

webapp-node-azureTest - Added some baby tests for manage